Refactor WooCommerceBundledBadge component to be generic receiving the details through props
