+++ client/components/theme-type-badge/style.scss
@@ -24,7 +24,7 @@
 
 .theme-type-badge-tooltip,
 .theme-type-badge-tooltip.premium-badge__popover.popover,
-.theme-type-badge-tooltip.woocommerce-bundled-badge__popover, {
+.theme-type-badge-tooltip.bundled-badge__popover {
 	.popover__inner {
 		background: #fff;
 		color: var(--color-neutral-50);
