+++ packages/components/src/theme-type-badge/index.ts
@@ -1,2 +1,2 @@
 export { default as PremiumBadge } from './premium-badge';
-export { default as WooCommerceBundledBadge } from './woocommerce-bundled-badge';
+export { default as BundledBadge } from './bundled-badge';
