+++ packages/design-preview/src/components/style.scss
@@ -211,7 +211,7 @@ $design-preview-sidebar-width: 311px;
 
 .design-preview__sidebar-badges {
 	.premium-badge,
-	.woocommerce-bundled-badge {
+	.bundled-badge {
 		letter-spacing: 0.2px;
 		margin: 0;
 	}
